Moved files into a logical package
