+++ base/uk.ac.stfc.isis.ibex.ui.motor/src/uk/ac/stfc/isis/ibex/ui/motor/displayoptions/ColourOption.java
@@ -17,7 +17,7 @@
  * http://opensource.org/licenses/eclipse-1.0.php
  */
 
-package uk.ac.stfc.isis.ibex.ui.motor;
+package uk.ac.stfc.isis.ibex.ui.motor.displayoptions;
 
 /**
  * Enum listing colour palette options.
