+++ base/uk.ac.stfc.isis.ibex.ui.motor/src/uk/ac/stfc/isis/ibex/ui/motor/views/TableOfMotorsView.java
@@ -23,7 +23,7 @@ import com.google.common.base.Strings;
 import uk.ac.stfc.isis.ibex.motor.Motor;
 import uk.ac.stfc.isis.ibex.motor.Motors;
 import uk.ac.stfc.isis.ibex.motor.internal.MotorsTable;
-import uk.ac.stfc.isis.ibex.ui.motor.DisplayPreferences;
+import uk.ac.stfc.isis.ibex.ui.motor.displayoptions.DisplayPreferences;
 
 /** A view that shows a collection of motors. */
 public class TableOfMotorsView extends ViewPart {
